api/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-CORS==3.0.10
python-dotenv==1.0.0
Werkzeug==2.2.3
MarkupSafe==2.1.2
Jinja2==3.1.2
itsdangerous==2.1.2
SQLAlchemy==2.0.0
pytz==2023.3.post1